move loggable to module folder
