+++ apps/server/src/modules/provisioning/loggable/index.ts
@@ -1,5 +1,8 @@
 export * from './user-for-group-not-found.loggable';
 export * from './school-for-group-not-found.loggable';
+export * from './bad-data.loggable-exception';
+export * from './school-name-required-loggable-exception';
+
 export * from './group-role-unknown.loggable';
 export { SchoolExternalToolCreatedLoggable } from './school-external-tool-created.loggable';
 export { FetchingPoliciesInfoFailedLoggable } from './fetching-policies-info-failed.loggable';
