+++ apps/server/src/modules/provisioning/strategy/oidc/service/schulconnex-school-provisioning.service.ts
@@ -5,7 +5,7 @@ import { LegacySchoolDo } from '@shared/domain/domainobject';
 import { FederalStateEntity, SchoolYearEntity } from '@shared/domain/entity';
 import { EntityId, SchoolFeature } from '@shared/domain/types';
 import { ExternalSchoolDto } from '../../../dto';
-import { SchoolNameRequiredLoggableException } from '../../loggable/school-name-required-loggable-exception';
+import { SchoolNameRequiredLoggableException } from '../../../loggable';
 
 @Injectable()
 export class SchulconnexSchoolProvisioningService {
